Ensure empty strings are set for plural messages (#31)
Previously it didn't work very well to set new plural messages. The
msgid and msgid_plural was set as expected. But msgstr[0,1] was not
visible in the output at all.

This PR fixes that by ensuring that there's always at least an empty
translations available for both singular and plural messages.

Fixes #30